requirements.txt
# PersonalSkin — פרויקט הגשה (Python / Flask)
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.0
WTForms>=3.1.0
email-validator>=2.1.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
openai>=1.40.0
markdown>=3.6
bleach>=6.1.0
